ABMC: Weakened the condition on fixed-point detection; fixed computation of state invariant
Previously the condition on fixed involved two steps of abstract transition relation on level n to be included in one step of this relation. This is unnecessary strong, it is sufficient if one step of this relation combined with one step of true transition relation is included in one step of this relation.